TBR: fix build break
BUG=

Review URL: https://codereview.chromium.org//25415002

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@28076 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/object_test.cc
@@ -1544,7 +1544,8 @@ TEST_CASE(EscapeSpecialCharactersOneByteString) {
       String::Handle(String::EscapeSpecialCharacters(str));
   EXPECT(escaped_str.Equals("a\\n\\f\\b\\t\\v\\r\\\\\\$z"));
 
-  const String& empty_str = String::Handle(OneByteString::New(0, Heap::kNew));
+  const String& empty_str = String::Handle(
+      OneByteString::New(static_cast<intptr_t>(0), Heap::kNew));
   const String& escaped_empty_str =
       String::Handle(String::EscapeSpecialCharacters(empty_str));
   EXPECT_EQ(empty_str.Length(), 0);
@@ -1591,7 +1592,8 @@ TEST_CASE(EscapeSpecialCharactersTwoByteString) {
       String::Handle(String::EscapeSpecialCharacters(str));
   EXPECT(escaped_str.Equals("a\\n\\f\\b\\t\\v\\r\\\\\\$z"));
 
-  const String& empty_str = String::Handle(TwoByteString::New(0, Heap::kNew));
+  const String& empty_str =
+      String::Handle(TwoByteString::New(static_cast<intptr_t>(0), Heap::kNew));
   const String& escaped_empty_str =
       String::Handle(String::EscapeSpecialCharacters(empty_str));
   EXPECT_EQ(empty_str.Length(), 0);
